pyproject.toml
[project]
name = "juniper-mist-mcp"
version = "0.1.0"
description = "MCP server for Juniper Mist API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.5.1",
    "mistapi>=0.61.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
